Add an abstract layer for analytics, fix logging change of snippet, fix logging cancelling, fix logging link following (#22012)
